+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Clean up extraneous error messages in extensions commands.
+* Adds breakpoint debugging for the Cloud Functions emulator using the `--inspect-functions` flag (#1360).